requirements.txt
attrs==22.1.0
azure-ai-anomalydetector==3.0.0b6
azure-core==1.26.4
azure-storage-blob==12.16.0
bcrypt==4.0.1
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
contourpy==1.0.7
crashtest==0.4.1
cryptography==37.0.4
cycler==0.11.0
distlib==0.3.6
docopt==0.6.2
dulwich==0.20.46
filelock==3.8.0
fonttools==4.39.4
html5lib==1.1
idna==3.4
isodate==0.6.1
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
kiwisolver==1.4.4
lockfile==0.12.2
matplotlib==3.7.1
more-itertools==8.14.0
msgpack==1.0.4
numpy==1.24.3
packaging==22.0
pandas==2.0.1
paramiko==3.1.0
pexpect==4.8.0
Pillow==9.5.0
pipenv==2022.12.19
pipreqs==0.4.13
pkginfo==1.9.2
platformdirs==2.5.2
plotly==5.14.1
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2023.3
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
storage==0.0.4.3
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
yarg==0.1.9

python-dotenv~=1.0.0